+++ LeanAndMeanCards/Plugin.cs
@@ -24,7 +24,7 @@ namespace LeanAndMeanCards
     {
         public const string ModId = "com.ljindustries.rounds.leanandmeancards";
         public const string ModName = "Lean and Mean Cards";
-        public const string Version = "1.1.0";
+        public const string Version = "1.1.1";
         public const string ModInitials = "LMC";
         public const string CardsMenuName = "LeanAndMeanCards";